# docs/swagger.yaml
basePath: /api/v1
definitions:
  apitypes.MassCalculationJSON:
    properties:
      creator_login:
        type: string
      date_create:
        type: string
      date_finish:
        type: string
      date_form:
        type: string
      id:
        type: integer
      moderator_login:
        type: string
      output_koef:
        type: number
      status:
        type: string
    type: object
  apitypes.ReactionCalculationJSON:
    properties:
      calculation_id:
        type: integer
      id:
        type: integer
      input_mass:
        type: number
      output_mass:
        type: number
      reaction_id:
        type: integer
    type: object
...
  apitypes.UserJSON:
    properties:
      is_moderator:
        type: boolean
      login:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для работы с реакциями
  title: Reactions API
  version: "1.0"
paths:
...
  /reaction-calculations/{mass_calculation_id}/{reaction_id}:
    delete:
      description: Удаляет связь реакции и расчета и возвращает обновлённый расчет.
      parameters:
      - description: ID расчета
        in: path
        name: mass_calculation_id
        required: true
        type: integer
      - description: ID реакции
        in: path
        name: reaction_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый расчет
          schema:
            $ref: '#/definitions/apitypes.MassCalculationJSON'
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещён
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить реакцию из расчета
      tags:
      - reaction-calculations
    put:
      consumes:
      - application/json
      description: Обновляет поле расчёта реакции (output mass).
      parameters:
      - description: ID расчета
        in: path
        name: mass_calculation_id
        required: true
        type: integer
      - description: ID реакции
        in: path
        name: reaction_id
        required: true
        type: integer
      - description: Новые значения расчёта
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apitypes.ReactionCalculationJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый расчёт
          schema:
            $ref: '#/definitions/apitypes.ReactionCalculationJSON'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещён
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить поле связи многие ко многим
      tags:
      - reaction-calculations
...
  /users/{login}/profile:
    get:
      description: Возвращает данные профиля (доступен только тот, чей UUID совпадает
        с user_id в токене).
      parameters:
      - description: Логин пользователя
        in: path
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя
          schema:
            $ref: '#/definitions/apitypes.UserJSON'
        "400":
          description: Проблема с авторизацией/получением user_id"}
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователи не совпадают
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить профиль пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет профиль пользователя (может делать только сам пользователь).
      parameters:
      - description: Логин пользователя
        in: path
        name: login
        required: true
        type: string
      - description: Новые данные профиля
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apitypes.UserJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый профиль
          schema:
            $ref: '#/definitions/apitypes.UserJSON'
        "400":
          description: Ошибка запроса или авторизации
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещён
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить профиль пользователя
      tags:
      - users
  /users/sign-in:
    post:
      consumes:
      - application/json
      description: Принимает логин/пароль, возвращает jwt-токен в формате {"token":"..."}.
      parameters:
      - description: Логин и пароль
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/apitypes.UserJSON'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход (получение токена)
      tags:
      - users
  /users/sign-out:
    post:
      description: Удаляет токен текущего пользователя из хранилища. Возвращает {"status":"signed_out"}.
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Проблема с получением user_id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка при удалении токена
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выход (удаление токена)
      tags:
      - users
  /users/sign-up:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя. Возвращает URL созданного ресурса
        в Location и тело созданного пользователя.
      parameters:
      - description: Параметры нового пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apitypes.UserJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/apitypes.UserJSON'
        "400":
          description: Ошибка валидации или входных данных
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"